+++ NEWS.md
@@ -4,6 +4,8 @@ NEW FUNCTIONS
 
 * `data_summary()`, to compute summary statistics of (grouped) data frames.
 
+* `value_at()`, to extract values at a specific position in a vector or factor.
+
 CHANGES
 
 * `data_modify()` gets three new arguments, `.at`, `.if` and `.modify`, to modify